refactor: restructure stringzy package to TypeScript
- Removed index.js and transformed the package structure to use TypeScript.
- Added TypeScript configuration (tsconfig.json) for project setup.
- Created separate TypeScript files for analyzing, formatting, transformations, and validations.
- Updated package.json to reflect new entry points and added build scripts.
- Added devDependencies for TypeScript and type definitions.
- Implemented functions for character counting, frequency analysis, complexity scoring, and various string transformations.
- Introduced validation functions for URL, email, date, and slug checks.
- Enhanced formatting functions for capitalization, number formatting, and phone formatting.
- Removed old JavaScript files and replaced them with TypeScript implementations.
