+++ .npmignore
@@ -1,1 +1,5 @@
 assets/
+src/
+
+tsconfig.json
+tsconfig.build.json
