+++ package.json
@@ -1,11 +1,14 @@
 {
   "name": "stringzy",
   "version": "2.2.1",
-  "main": "index.js",
+  "main": "dist/index.js",
   "type": "module",
-  "types": "index.d.ts",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "lint": "eslint . --ext .js,.ts",
+    "prepublishOnly": "npm run build"
   },
   "author": "Samarth Ruia",
   "readmeFilename": "README.md",
@@ -41,5 +44,9 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Samarth2190/stringzy.git"
+  },
+  "devDependencies": {
+    "@types/node": "^24.0.0",
+    "typescript": "^5.8.3"
   }
 }
