+++ src/analyzing/characterCount.ts
@@ -1,0 +1,7 @@
+export function characterCount(str: string): number {
+    if (typeof str !== "string") {
+        throw new TypeError("Input must be a string");
+    }
+
+    return str.length;
+}
