+++ src/analyzing/index.ts
@@ -1,0 +1,5 @@
+export { characterCount } from "./characterCount";
+export { characterFrequency } from "./characterFrequency";
+export { complexity } from "./complexity";
+export { wordCount } from "./wordCount";
+
