+++ src/formatting/index.ts
@@ -1,0 +1,3 @@
+export { capitalize } from "./capitalize";
+export { formatNumber } from "./number";
+export { formatPhone } from "./phone";
