+++ src/transformations/kebabCase.ts
@@ -1,0 +1,11 @@
+export function kebabCase(text: string): string {
+    if (text == null) return '';
+    return text
+        .trim()
+        .replace(/[\s_]+/g, '-')
+        .replace(/([a-z])([A-Z])/g, '$1-$2')
+        .replace(/[^\w-]/g, '-')
+        .toLowerCase()
+        .replace(/-+/g, '-')
+        .replace(/^-+|-+$/g, '');
+}
