+++ src/transformations/snakeCase.ts
@@ -1,0 +1,11 @@
+export function snakeCase(text: string): string {
+    if (text == null) return '';
+    return text
+        .trim()
+        .replace(/[\s-]+/g, '_')
+        .replace(/([a-z])([A-Z])/g, '$1_$2')
+        .replace(/[^\w_]/g, '_')
+        .toLowerCase()
+        .replace(/_+/g, '_')
+        .replace(/^_+|_+$/g, '');
+}
