+++ src/transformations/titleCase.ts
@@ -1,0 +1,10 @@
+export function titleCase(text: string): string {
+    if (text == null) return '';
+    return text
+        .trim()
+        .toLowerCase()
+        .replace(/([^\w\s]|_)/g, ' ')
+        .replace(/\s+/g, ' ')
+        .replace(/(^|\s)(.)/g, (_, prefix, character) => prefix + character.toUpperCase())
+        .replace(/\s/g, ' ');
+}
