+++ src/validations/index.ts
@@ -1,0 +1,6 @@
+export { isDate } from "./isDate";
+export { isEmail } from "./isEmail";
+export { isEmpty } from "./isEmpty";
+export { isSlug } from "./isSlug";
+export { isURL } from "./isURL";
+
