+++ src/validations/isEmail.ts
@@ -1,0 +1,4 @@
+export function isEmail(str: string): boolean {
+  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+  return emailRegex.test(str);
+}
