+++ src/validations/isEmpty.ts
@@ -1,0 +1,3 @@
+export function isEmpty(str: string): boolean {
+  return str.trim().length === 0;
+}
