+++ src/validations/isURL.ts
@@ -1,0 +1,8 @@
+export function isURL(str: string): boolean {
+  try {
+    new URL(str);
+    return true;
+  } catch {
+    return false;
+  }
+}
